+++ vim_rs/src/types/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "defaults")]
+pub mod defaults;
 pub mod enums;
 pub mod structs;
 pub mod traits;